Change Recipe classes from being nested inside of RecipeDefinition into their own files and a new Recipes namespace
