+++ src/AWS.Deploy.Common/Recipes/DeploymentBundleTypes.cs
@@ -1,0 +1,11 @@
+﻿// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+namespace AWS.Deploy.Common.Recipes
+{
+    public enum DeploymentBundleTypes
+    {
+        DotnetPublishZipFile,
+        Container
+    }
+}
