+++ src/AWS.Deploy.Common/Recipes/DeploymentTypes.cs
@@ -1,0 +1,10 @@
+﻿// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+namespace AWS.Deploy.Common.Recipes
+{
+    public enum DeploymentTypes
+    {
+        CdkProject
+    }
+}
