+++ src/AWS.Deploy.Common/Recipes/OptionSettingValueType.cs
@@ -1,0 +1,12 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+namespace AWS.Deploy.Common.Recipes
+{
+    public enum OptionSettingValueType
+    {
+        String,
+        Int,
+        Bool
+    };
+}
