+++ src/AWS.Deploy.Common/Recipes/PropertyDependency.cs
@@ -1,0 +1,14 @@
+﻿// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+namespace AWS.Deploy.Common.Recipes
+{
+    /// <summary>
+    /// Container for property dependencies
+    /// </summary>
+    public class PropertyDependency
+    {
+        public string Id { get; set; }
+        public string Value { get; set; }
+    }
+}
